split sidebar to a file
